+++ src/utils/MissionUtils.js
@@ -32,33 +32,6 @@ const getUserNumber = (computer, question) => {
   });
 };
 
-const validateUserNumber = (userInput) => {
-  const isBetweenOneToNine = validateOneToNine(userInput);
-  const isCorrectLength = validateLength(userInput, 3);
-  const isNotOverlapped = validateOverlapped(userInput);
-
-  if (isBetweenOneToNine && isCorrectLength && isNotOverlapped) {
-    return true;
-  } else {
-    throw new Error("에러 발생 후 종료");
-  }
-};
-
-const validateLength = (target, setLength) => {
-  let validLength = target.length === setLength;
-  return validLength;
-};
-
-const validateOverlapped = (target) => {
-  const setTarget = new Set(target);
-  let validOverlapped = target.length === setTarget.size;
-  return validOverlapped;
-};
-
-const validateOneToNine = (target) => {
-  return NUMBER_ONE_TO_NINE.test(target);
-};
-
 const countStrikeAndBall = (computers, user) => {
   const users = useNumberToArray(user);
 
@@ -139,7 +112,6 @@ const retryOrEnd = () => {
 module.exports = {
   getUserNumber,
   setUserInput,
-  validateUserNumber,
   countStrikeAndBall,
   isUserWin,
   printGameResultMessage,
